Refactor SDK to minimize network calls and properly create interface boundaries (#94)
Given that we're in SDK land, I think that the default should be to use
the MarketMetadata object as opposed to the MarketHeader object. There
were previously a lot of areas where we seem to be indecisive on which
object to use when grabbing market metadata, and now I think every case
is disambiguated.

There should be 0 interface changes that arise from this PR.


Changes:
- Add unit conversion helper to MarketMetadata struct
- Still enable SDK users to look up conversion by market
- No longer force users to call add market to a mutable SDK prior to
using functions like `get_market_orderbook`

Notes:
- The code now makes a few more copies than before (need to load
MarketMetadata into a struct). This itself is super small and should
have 0 impact on code perf
- I added back helper conversions that were removed from a previous PR